# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/accounts/account": {
            "get": {
                "tags": [
                    "Accounts",
                    "Unfinished"
                ],
                "summary": "Get Account",
                "operationId": "get_account_accounts_account_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": [
                            "read:accounts"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Account Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Account"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/accounts/club/account": {
            "post": {
                "tags": [
                    "Accounts",
                    "Unfinished",
                    "Clubs"
                ],
                "summary": "Create Account",
                "operationId": "create_account_accounts_club_account_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Account"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Account"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": [
                            "write:accounts"
                        ]
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Accounts",
                    "Unfinished",
                    "Clubs"
                ],
                "summary": "Update Club Account",
                "operationId": "update_club_account_accounts_club_account_patch",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Account"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Account"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/accounts/accounts": {
            "get": {
                "tags": [
                    "Accounts",
                    "Unfinished"
                ],
                "summary": "Get Accounts",
                "operationId": "get_accounts_accounts_accounts_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": [
                            "read:accounts"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "title": "Range starting ID to get"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "title": "Amount of resources to fetch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Account"
                                    },
                                    "title": "Response Get Accounts Accounts Accounts Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/accounts/user/account": {
            "patch": {
                "tags": [
                    "Accounts",
                    "Unfinished",
                    "Users"
                ],
                "summary": "Update User Account",
                "operationId": "update_user_account_accounts_user_account_patch",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Account"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Account"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/accounts/account/me": {
            "get": {
                "tags": [
                    "Accounts",
                    "Unfinished",
                    "Users"
                ],
                "summary": "Get User Account",
                "operationId": "get_user_account_accounts_account_me_get",
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Account"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/accounts/user/account/{user_id}": {
            "get": {
                "tags": [
                    "Accounts",
                    "Unfinished",
                    "Users"
                ],
                "summary": "Get User Account",
                "operationId": "get_user_account_accounts_user_account__user_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": [
                            "read:accounts"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Id"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Account"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/accounts/club/account/{account_id}/": {
            "get": {
                "tags": [
                    "Accounts",
                    "Unfinished",
                    "Clubs"
                ],
                "summary": "Get Club Account",
                "operationId": "get_club_account_accounts_club_account__account_id___get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Account Id"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Account"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/accounts/club/account/id/{account_id}": {
            "get": {
                "tags": [
                    "Accounts",
                    "Unfinished",
                    "Clubs"
                ],
                "summary": "Get Account With Club",
                "operationId": "get_account_with_club_accounts_club_account_id__account_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Account Id"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Club-Output"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/accounts/club/accounts": {
            "get": {
                "tags": [
                    "Accounts",
                    "Unfinished",
                    "Clubs"
                ],
                "summary": "Get Club Accounts",
                "operationId": "get_club_accounts_accounts_club_accounts_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "club_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "title": "Club ID off accounts"
                        }
                    },
                    {
                        "name": "start",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "title": "Range starting ID to get"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "title": "Amount of resources to fetch"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Account"
                                    },
                                    "title": "Response Get Club Accounts Accounts Club Accounts Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/accounts/account/top-up": {
            "post": {
                "tags": [
                    "Accounts",
                    "Unfinished",
                    "Clubs"
                ],
                "summary": "Club Top Up",
                "operationId": "club_top_up_accounts_account_top_up_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TopUp"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Account": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "uid": {
                        "type": "string",
                        "title": "Uid"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "entity_type": {
                        "type": "string",
                        "title": "Entity Type"
                    },
                    "entity_id": {
                        "type": "integer",
                        "title": "Entity Id"
                    },
                    "discount_rate": {
                        "type": "integer",
                        "title": "Discount Rate",
                        "max_value": 100,
                        "min_value": 0
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "balance": {
                        "type": "number",
                        "title": "Balance"
                    },
                    "class_name": {
                        "type": "string",
                        "title": "Class Name",
                        "default": "ACCOUNT"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "uid",
                    "name",
                    "entity_type",
                    "entity_id",
                    "discount_rate",
                    "status",
                    "balance"
                ],
                "title": "Account"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "TopUp": {
                "properties": {
                    "card_id": {
                        "type": "integer",
                        "title": "Card Id"
                    },
                    "account_id": {
                        "type": "integer",
                        "title": "Account Id"
                    },
                    "amount": {
                        "type": "number",
                        "title": "Amount"
                    },
                    "user_password": {
                        "type": "string",
                        "title": "User Password"
                    }
                },
                "type": "object",
                "required": [
                    "card_id",
                    "account_id",
                    "amount",
                    "user_password"
                ],
                "title": "TopUp"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "token"
                    }
                }
            }
        }
    }
}